# Desktop/彼雅特github/bityacht-exchange-api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errpkg.Code:
    enum:
    - 4000
    - 4001
    - 4002
    - 4003
    - 4004
    - 4005
    - 4006
    - 4007
    - 4008
    - 4009
    - 4010
    - 4011
    - 4012
    - 4013
    - 4014
    - 4015
    - 4016
    - 4017
    - 4018
    - 4019
    - 4020
    - 4021
    - 4022
    - 4023
    - 4024
    - 4025
    - 4026
    - 4027
    - 4028
    - 4029
    - 4030
    - 4031
    - 4032
    - 4033
    - 4034
    - 4035
    - 4036
    - 4037
    - 4038
    - 4039
    - 4040
    - 4041
    - 4042
    - 4043
    - 4044
    - 5000
    - 5001
    - 5002
    - 5003
    - 5004
    - 5005
    - 5006
    - 5007
    - 5008
    - 5009
    - 5010
    - 5011
    - 5012
    - 5013
    - 5014
    - 5015
    - 5016
    - 5017
    - 5018
    - 5019
    - 5020
    - 5021
    - 5022
    - 5023
    - 5024
    - 5025
    - 5026
    - 5027
    - 5028
    - 5029
    - 5030
    - 5031
    - 5032
    - 5033
    - 5034
    - 5035
    - 5036
    - 5037
    - 5038
    - 5039
    - 5040
    - 5041
    - 9993
    - 9994
    - 9995
    - 9996
    - 9997
    - 9998
    - 9999
    type: integer
    x-enum-comments:
      CodeBadErrorType: Use general error instead of *Error
    x-enum-varnames:
    - CodeBindBody
    - CodeUnauthorized
    - CodeJWTBadSigning
    - CodeJWTInvalid
    - CodeJWTRevoked
    - CodeBadAuthorizationToken
    - CodePermissionDenied
    - CodeAccountDuplicated
    - CodeBadParam
    - CodeRecordNotFound
    - CodeRecordNoChange
    - CodeBadBody
    - CodeAccountNotAvailable
    - CodeInsufficientFunds
    - CodeBadQuery
    - CodeFundDeleted
    - CodeTooManyRequests
    - CodeBadPasswordStrength
    - CodeOverResetPasswordLimit
    - CodeBadUploadedFileType
    - CodeFileNotFound
    - CodeEmailAlreadyVerified
    - CodeBadVerificationCode
    - CodeVerificationCodeExpired
    - CodeBadMobileBarcodeFormat
    - CodeBadInviteCode
    - CodeImageOverSize
    - CodeBadDataURIImageFormat
    - CodeBadImageData
    - CodeBadNationalID
    - CodeNationalIDDuplicated
    - CodeBadAction
    - CodeInsufficientBalance
    - CodeBadAmount
    - CodeTransactionPairNotFound
    - CodeCurrencyNotFound
    - CodeBadCalcOfTransaction
    - CodeBadTransactionPairStatus
    - CodeTemporaryForbidden
    - CodeBadCSVContent
    - CodeMobileBarcodeNotExist
    - CodeBadCryptocurrencyAddress
    - CodePhoneNumberDuplicated
    - CodeOverWithdrawalWhitelistLimit
    - CodeTooManyFiles
    - CodeSql
    - CodeJWTIssueToken
    - CodeGetClaims
    - CodeBadClaimsType
    - CodeSendEmail
    - CodeJWTBadPayload
    - CodeEncryption
    - CodeGormScan
    - CodeGormValue
    - CodeNewHTTPRequest
    - CodeDoHTTPRequest
    - CodeReadHTTPResponse
    - CodeSendSMS
    - CodeJSONMarshal
    - CodeJSONUnmarshal
    - CodeBadAmountOfFundsInTransactions
    - CodeBadRecord
    - CodeSaveUploadedFile
    - CodeRedis
    - CodeRedisBadScript
    - CodeCallBinanceAPI
    - CodeParseURL
    - CodeCallMaxAPI
    - CodeJSONDecode
    - CodeParseIP
    - CodeLookUpMMDB
    - CodeCallEZReceiptAPI
    - CodeLoginEZReceiptAPI
    - CodeBadParamEZReceiptAPI
    - CodeWriteCSV
    - CodeCallKryptoGOAPI
    - CodeExecuteTemplate
    - CodeUpdateCache
    - CodeCallCybavoAPI
    - CodeCybavoWalletNotFound
    - CodeWalletAddressNotGen
    - CodeWalletAddressAlreadySet
    - CodeGenQRCode
    - CodeAddressDeploying
    - CodeMemoryCacheError
    - CodeBadBase32String
    - CodeGenerateHOTP
    - CodeBadErrorType
    - CodeFailedToGenerateID
    - CodeScheduleJobKeyDuplicated
    - CodeBadCoding
    - CodeNotInit
    - CodeNotImplement
    - CodeUnknown
  errpkg.JsonError:
    properties:
      code:
        $ref: '#/definitions/errpkg.Code'
      data: {}
      error:
        type: string
    required:
    - code
    type: object
...
  modelpkg.GetResponse:
    properties:
      data: {}
      page:
        description: 'Default: 1'
        type: integer
      pageSize:
        description: 'Default: 25'
        type: integer
    required:
    - data
    type: object
...
  risks.CreateRequest:
    properties:
      detail:
        type: string
      factor:
        type: string
      score:
        type: integer
      subFactor:
        type: string
    required:
    - detail
    - factor
    - score
    - subFactor
    type: object
  risks.Risk:
    properties:
      createdAt:
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
      detail:
        type: string
      factor:
        type: string
      id:
        type: integer
      score:
        type: integer
      subFactor:
        type: string
    type: object
  risks.UpdateRequest:
    properties:
      detail:
        type: string
      factor:
        type: string
      score:
        type: integer
      subFactor:
        type: string
    type: object
...
  suspicioustransactions.SuspiciousTX:
    properties:
      createdAt:
        description: 掃描時間
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
      dedicatedReviewResult:
        description: |-
          狀態 (= 專責審查結果)
          * 1: 待審核
          * 2: 通過
          * 3: 駁回
        type: integer
      email:
        description: E-Mail
        type: string
      firstName:
        description: 名
        type: string
      id:
        description: 案件單號
        type: integer
      lastName:
        description: 姓
        type: string
      orderID:
        description: 訂單編號
        type: string
      type:
        allOf:
        - $ref: '#/definitions/suspicioustransactionspkg.Type'
        description: |-
          可疑樣態
          * 1: 多筆提領態樣
          * 2: 多筆同額態樣
          * 3: 同一出金地址態樣
          * 4: 迅速轉出態樣
          * 5: 迅速買賣態樣
          * 6: 小額接收大額轉出態樣
      usersID:
        description: UID
        type: integer
    type: object
  suspicioustransactions.SuspiciousTXDetail:
    properties:
      annualIncome:
        description: 年收入
        type: string
      countriesCode:
        description: 國籍
        type: string
      createdAt:
        description: 掃描時間
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
      dedicatedReviewComment:
        description: 專責審查(備註)
        type: string
      dedicatedReviewResult:
        description: |-
          狀態 (= 專責審查結果)
          * 1: 待審核
          * 2: 通過
          * 3: 駁回
        type: integer
      dualNationalityCode:
        description: 雙重國籍
        type: string
      email:
        description: E-Mail
        type: string
      firstName:
        description: 名
        type: string
      fundsSources:
        description: 資金來源
        type: string
      id:
        description: 案件單號
        type: integer
      industrialClassificationsID:
        description: 行業別
        type: integer
      informationReviewComment:
        description: 資訊審核(備註)
        type: string
      informationReviewFiles:
        description: 資訊審核(檔案)
        items:
          type: string
        type: array
      informations:
        allOf:
        - $ref: '#/definitions/suspicioustransactionspkg.Information'
        description: 交易資訊
      lastName:
        description: 姓
        type: string
      nationalID:
        description: 身份證字號
        type: string
      orderID:
        description: 訂單編號
        type: string
      phone:
        description: 手機
        type: string
      purposeOfUse:
        description: 使用目的
        type: string
      registerAt:
        description: 註冊時間
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
      registerIP:
        description: 註冊IP
        type: string
      reportMJIBAt:
        description: 呈報調查局日期
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
      riskReviewFiles:
        description: 風控審查(檔案)
        items:
          type: string
        type: array
      riskReviewResult:
        description: |-
          風控審查(結果)
          * 1: 待審核
          * 2: 非可疑交易
          * 3: 有可疑交易風險
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/suspicioustransactionspkg.Type'
        description: |-
          可疑樣態
          * 1: 多筆提領態樣
          * 2: 多筆同額態樣
          * 3: 同一出金地址態樣
          * 4: 迅速轉出態樣
          * 5: 迅速買賣態樣
          * 6: 小額接收大額轉出態樣
      usersID:
        description: UID
        type: integer
    type: object
...
  suspicioustransactions.UpdateRequest:
    properties:
      comment:
        description: 備註 (資訊審核 or 專責審查)
        type: string
      dedicatedReviewResult:
        description: |-
          專責審查結果
          * 1: 待審核
          * 2: 通過
          * 3: 駁回
        type: integer
      reportMJIBAt:
        description: 呈報調查局日期
        format: dateTime(YYYY/MM/DD HH:mm:SS)
        type: string
      riskReviewResult:
        description: |-
          風控審查結果
          * 1: 待審核
          * 2: 非可疑交易
          * 3: 有可疑交易風險
        type: integer
      type:
        allOf:
        - $ref: '#/definitions/suspicioustransactions.UpdateType'
        description: |-
          更新類型
          * 1: 資訊審核 (所需欄位：備註)
          * 2: 風控審查結果 (所需欄位：風控審查結果)
          * 3: 專責審查 (所需欄位：專責審查結果、備註)
          * 4: 呈報調查局日期 (所需欄位：呈報調查局日期)
        maximum: 4
        minimum: 1
    required:
    - type
    type: object
  suspicioustransactions.UpdateType:
    enum:
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - UpdateTypeInformationReviewComment
    - UpdateTypeRiskReviewResult
    - UpdateTypeDedicatedReview
    - UpdateTypeReportMJIBAt
  suspicioustransactionspkg.Information:
    properties:
      loginAt:
        type: string
      spotTransfers:
        items:
          $ref: '#/definitions/usersspottransfers.Transfer'
        type: array
      transactions:
        items:
          $ref: '#/definitions/userstransactions.TransactionForManager'
        type: array
    type: object
  suspicioustransactionspkg.Type:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    type: integer
    x-enum-varnames:
    - TypeMultipleWithdrawal
    - TypeMultipleSameAmount
    - TypeMultipleSameWithdrawalAddress
    - TypeQuicklyWithdraw
    - TypeQuicklyTrade
    - TypeSmallInBigOut
  transactionpairs.Status:
    enum:
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - StatusDisable
    - StatusEnable
  transactionpairs.TransactionPair:
    properties:
      baseCurrenciesDecimalPrecision:
        type: integer
      baseCurrenciesSymbol:
        type: string
      handlingChargeRate:
        format: Decimal(27, 9)
        type: string
      quoteCurrenciesDecimalPrecision:
        type: integer
      quoteCurrenciesSymbol:
        type: string
      spreadsOfBuy:
        format: Decimal(27, 9)
        type: string
      spreadsOfsell:
        format: Decimal(27, 9)
        type: string
      status:
        $ref: '#/definitions/transactionpairs.Status'
    type: object
  transactionpairs.UpdateRequest:
    properties:
      baseCurrenciesSymbol:
        type: string
      handlingChargeRate:
        format: Decimal(27, 9)
        type: string
      quoteCurrenciesSymbol:
        type: string
      spreadsOfBuy:
        format: Decimal(27, 9)
        type: string
      spreadsOfsell:
        format: Decimal(27, 9)
        type: string
      status:
        $ref: '#/definitions/transactionpairs.Status'
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: This is a RESTful API documentation of BitYacht Exchange Backend.
  title: BitYacht Exchange API Doc
  version: 0.1.0
paths:
...
  /admin/kycs/risks:
    get:
      description: 取得所有風險設定
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/risks.Risk'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 取得所有風險設定
      tags:
      - Admin-kycs
    post:
      consumes:
      - application/json
      description: 新增一個風險設定
      parameters:
      - description: Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/risks.CreateRequest'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 新增一個風險設定
      tags:
      - Admin-kycs
  /admin/kycs/risks/{id}:
    delete:
      description: 刪除一個風險設定, 會同時刪除所有關聯的風險設定
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 刪除一個風險設定
      tags:
      - Admin-kycs
    patch:
      consumes:
      - application/json
      description: 修改一個風險設定
      parameters:
      - description: Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/risks.UpdateRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 修改一個風險設定
      tags:
      - Admin-kycs
...
  /admin/settings/transactionpairs:
    get:
      description: 取得交易對列表
      parameters:
      - description: 'Default: 1'
        in: query
        name: page
        type: integer
      - description: 'Default: 25'
        in: query
        name: pageSize
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/modelpkg.GetResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/transactionpairs.TransactionPair'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 取得交易對列表
      tags:
      - Admin-settings
    patch:
      description: 編輯交易對
      parameters:
      - description: Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/transactionpairs.UpdateRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 編輯交易對
      tags:
      - Admin-settings
...
  /admin/suspicious-txs:
    get:
      description: 取得可疑交易列表
      parameters:
      - description: 'Default: 1'
        in: query
        name: page
        type: integer
      - description: 'Default: 25'
        in: query
        name: pageSize
        type: integer
      - description: 'Default: '''''
        in: query
        name: search
        type: string
      - description: |-
          狀態
          * 0: 全部
          * 1: 待審核
          * 2: 通過
          * 3: 駁回
        in: query
        maximum: 3
        minimum: 0
        name: dedicatedReviewResult
        type: integer
      - description: 掃描時間（結束）
        format: date(YYYY/MM/DD)
        in: query
        name: endAt
        type: string
      - description: 掃描時間（開始）
        format: date(YYYY/MM/DD)
        in: query
        name: startAt
        type: string
      - description: |-
          可疑樣態
          * 0: 全部
          * 1: 多筆提領態樣
          * 2: 多筆同額態樣
          * 3: 同一出金地址態樣
          * 4: 迅速轉出態樣
          * 5: 迅速買賣態樣
          * 6: 小額接收大額轉出態樣
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        maximum: 6
        minimum: 0
        name: type
        type: integer
        x-enum-varnames:
        - TypeMultipleWithdrawal
        - TypeMultipleSameAmount
        - TypeMultipleSameWithdrawalAddress
        - TypeQuicklyWithdraw
        - TypeQuicklyTrade
        - TypeSmallInBigOut
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/modelpkg.GetResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/suspicioustransactions.SuspiciousTX'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 取得可疑交易列表
      tags:
      - Admin-SuspiciousTxs
  /admin/suspicious-txs/:id:
    get:
      description: 取得可疑交易詳細資訊
      parameters:
      - description: Suspicious Transactions ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/suspicioustransactions.SuspiciousTXDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 取得可疑交易詳細資訊
      tags:
      - Admin-SuspiciousTxs
    patch:
      description: 更新可疑交易資訊
      parameters:
      - description: Suspicious Transactions ID
        in: path
        name: id
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/suspicioustransactions.UpdateRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      security:
      - BearerAuth: []
      summary: 更新可疑交易資訊
      tags:
      - Admin-SuspiciousTxs
  /admin/suspicious-txs/:id/file:
    delete:
      description: 刪除可疑交易相關檔案
      parameters:
      - description: Suspicious Transactions ID
        in: path
        name: id
        required: true
        type: string
      - description: |-
          檔案類型
          * 1: 資訊審核
          * 2: 風控審查
        enum:
        - 1
        - 2
        in: query
        maximum: 2
        minimum: 1
        name: fileType
        required: true
        type: integer
        x-enum-varnames:
        - UpdateFilesTypeInformationReview
        - UpdateFilesTypeRiskReview
      - description: 檔案名稱
        in: query
        name: filename
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      summary: 刪除可疑交易相關檔案
      tags:
      - Admin-SuspiciousTxs
    get:
      description: 下載可疑交易相關檔案
      parameters:
      - description: Suspicious Transactions ID
        in: path
        name: id
        required: true
        type: string
      - description: |-
          檔案類型
          * 1: 資訊審核
          * 2: 風控審查
        enum:
        - 1
        - 2
        in: query
        maximum: 2
        minimum: 1
        name: fileType
        required: true
        type: integer
        x-enum-varnames:
        - UpdateFilesTypeInformationReview
        - UpdateFilesTypeRiskReview
      - description: 檔案名稱
        in: query
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      summary: 下載可疑交易相關檔案
      tags:
      - Admin-SuspiciousTxs
    post:
      consumes:
      - multipart/form-data
      description: 上傳可疑交易相關檔案
      parameters:
      - description: Suspicious Transactions ID
        in: path
        name: id
        required: true
        type: string
      - description: file
        in: formData
        name: file
        type: file
      - description: |-
          上傳類型
          * 1: 資訊審核
          * 2: 風控審查
        enum:
        - 1
        - 2
        in: formData
        maximum: 2
        minimum: 1
        name: uploadType
        required: true
        type: integer
        x-enum-varnames:
        - UpdateFilesTypeInformationReview
        - UpdateFilesTypeRiskReview
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      summary: 上傳可疑交易相關檔案
      tags:
      - Admin-SuspiciousTxs
  /admin/suspicious-txs/export-csv:
    get:
      consumes:
      - application/json
      description: 匯出可疑交易列表
      parameters:
      - collectionFormat: multi
        description: |-
          狀態 (Empty = All)
          * 1: 待審核
          * 2: 通過
          * 3: 駁回
        in: query
        items:
          type: integer
        name: dedicatedReviewResults
        type: array
      - description: 掃描時間（結束）
        format: date(YYYY/MM/DD)
        in: query
        name: endAt
        type: string
      - description: 掃描時間（開始）
        format: date(YYYY/MM/DD)
        in: query
        name: startAt
        type: string
      - collectionFormat: multi
        description: |-
          可疑樣態 (Empty = All)
          * 1: 多筆提領態樣
          * 2: 多筆同額態樣
          * 3: 同一出金地址態樣
          * 4: 迅速轉出態樣
          * 5: 迅速買賣態樣
          * 6: 小額接收大額轉出態樣
        in: query
        items:
          enum:
          - 1
          - 2
          - 3
          - 4
          - 5
          - 6
          type: integer
        name: types
        type: array
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errpkg.JsonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errpkg.JsonError'
      summary: 匯出可疑交易列表
      tags:
      - Admin-SuspiciousTxs
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"